Refactor: add wrapper components as `Board`
